Add matches to window.c
From PartyPlanner64/mp3.

+++ include/game/window.h
@@ -26,9 +26,12 @@ typedef struct TextWindow {
     /* 0x02B */ char pad2B[1];
     /* 0x02C */ s16 unk_2C;
     /* 0x02E */ s16 unk_2E;
-    /* 0x030 */ char pad30[4];                      /* maybe part of unk_2E[3]? */
+    /* 0x030 */ s8 unk_30;                      /* maybe part of unk_2E[3]? */
+    /* 0x031 */ s8 unk_31;
+    /* 0x032 */ s16 unk_32;
     /* 0x034 */ s16 unk_34;
-    /* 0x036 */ char pad36[6];                      /* maybe part of unk_34[4]? */
+    /* 0x036 */ char pad36[2];                      /* maybe part of unk_34[4]? */
+    /* 0x038 */ u32 unk_38;
     /* 0x03C */ u16 unk3C;
     /* 0x03E */ u16 unk3E;
     /* 0x040 */ char pad40[8];                      /* maybe part of unk3E[5]? */
@@ -61,7 +64,7 @@ void func_8005B43C_5C03C(s16, s32, s32, s32);
 void func_8005B6BC_5C2BC(s16 win_id, u32 arg1, s8 arg2);
 void func_8005BDFC_5C9FC(s16, s32);
 void func_8005BE30_5CA30(s16, s32);
-void func_8005BEE0_5CAE0(s16, s32);
+void func_8005BEE0_5CAE0(s16, s16);
 void func_8005C02C_5CC2C(s16, s32);
 void func_8005C060_5CC60(s16, s32, s32, s32, s32);
 void func_8005D294_5DE94(s16);
@@ -71,7 +74,7 @@ void func_8005F524_60124(void);
 void func_8005FBF8_607F8(s16, s32, s32, s32);
 void func_8005FE90_60A90(s16);
 void func_8005FFA8_60BA8(s16);
-void func_8006022C_60E2C(u32 mesg, s32);
+void func_8006022C_60E2C(u32 mesg, s16);
 void func_80060394_60F94(s32, s16*, s32);
 void func_80061388_61F88(s16);
 void func_80060144_60D44(s16 obj);
